requirements.txt
gradio>=3.50.0
paho-mqtt>=1.6.1
pandas>=2.0.0
pymongo>=4.5.0
prefect>=2.13.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
plotly>=5.18.0
asyncpg>=0.29.0
prometheus-client>=0.19.0
motor>=3.3.0
dbt-core>=1.7.0
dbt-postgres>=1.7.0